// client/src/app-constants/navigation.constant.ts
export enum ROUTES {
  HOME_PAGE = "/",
  PHONES = "/phones",
  PHONE_SLUG = "/phones/:slug",
  PHONE_REVIEWS = "phones/:slug/reviews",
  LAPTOPS = "/laptops",
  ORDER_HISTORY = "/order-history",
  CART = "/cart",
  // Auth
  LOGIN = "/login",
  SIGNUP = "/signup",
  FORGOTPASSWORD = "/forgotpassword",
  RESETPASSWORD = "/resetPassword",
  // firms
  // nested route
  SLUG = ":slug",
  PRODUCT_REVIEW = ":slug/reviews",
  PERSONAL_INFORMATION = "personal-information",
  IPHONE_13_SERRIES = "iphone-13-series",
  //
}